Update tax calculation feature/story-6
- Able to read formatted .csv file and calculate tax
- Update tax refund calculation
