+++ upload-csv/taxes.csv
@@ -1,4 +1,4 @@
-totalIncome,a,donation
+totalIncome,wht,donation
 500000,0,0
 600000,40000,20000
-750000,50000,15000,0
+750000,50000,15000
